Update AppView.php
Decided to extend to the View class of Cake itself because when testing I am getting errors. That means we depend on something insecure; wich is right, because there is a change that the user won't have its own View-class.

In future a callback could be used to keep changing the View-class.